+++ apps/code-examples/src/app/code-examples/lookup/lookup/add-item/lookup-demo-person.ts
@@ -1,0 +1,4 @@
+export interface LookupDemoPerson {
+  id: string;
+  name: string;
+}