+++ apps/e2e/pages-storybook-e2e/src/e2e/page-layout-blocks.component.cy.ts
@@ -15,6 +15,18 @@ describe(`pages-storybook`, () => {
           .screenshot(`${ID}-${theme}`)
           .percySnapshot(`${ID}-${theme}`);
       });
+
+      it('should render the component on mobile', () => {
+        cy.viewport(E2eVariations.MOBILE_WIDTHS[0], 800);
+
+        cy.get('app-blocks-page sky-page')
+          .should('exist')
+          .should('be.visible')
+          .screenshot(`${ID}-${theme}-mobile`)
+          .percySnapshot(`${ID}-${theme}-mobile`, {
+            widths: E2eVariations.MOBILE_WIDTHS,
+          });
+      });
     });
   });
 });